+++ docs/changelog.md
@@ -2,6 +2,7 @@
 ## v0.17.x
 - Consolidates the scope/binoculars field of view function into a general optic fov function (which means it also supports cameras now)
 - Adds models for comets and satellites and supports data formats from MPC and CelesTrak
+- Adds font name and weight style properties to legend labels
 
 
 ## v0.16.x